Addressed a show/hide animation bug.
The original code base contained a bug where if the user quickly swiped
up and down repeatedly on the AbsListView that a FAB was listening to,
the FAB would lose track of it’s original Y position and would slowly
scroll off screen never to be seen again. This fix has the FAB make a
note of it’s displayed Y position after it is assigned in onLayout()
and use it when animating instead of it’s current Y position.